+++ common/constants.ts
@@ -346,3 +346,4 @@ export const UI_TOAST_COLOR = {
   WARNING: 'warning',
   DANGER: 'danger',
 };
+